backend/src/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
gunicorn==21.2.0
openai==1.17.1
anthropic==0.25.6
azure-identity==1.16.0
python-dotenv # Pinned due to docker-in-docker issue: https://github.com/devcontainers/features/issues/616
environs==11.0.0
langfuse==2.28.2
asyncio==3.4.3
boto3==1.34.111
python-magic==0.4.27
python-multipart==0.0.6
PyPDF2==3.0.1
pandas==2.2.2
openpyxl==3.1.3
redis==5.0.6
auth0-python==4.7.1
cryptography==42.0.8
sse-starlette==2.1.2